+++ src/main/java/net/zaltren/firmabridge/proxy/CommonProxy.java
@@ -3,6 +3,7 @@ package net.zaltren.firmabridge.proxy;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+import net.minecraftforge.fml.common.registry.GameRegistry;
 import net.zaltren.firmabridge.FirmaBridge;
 import net.zaltren.firmabridge.compat.MaterialBridgeHandler;
 import net.zaltren.firmabridge.compat.OreQualityHandler;
@@ -12,6 +13,7 @@ import net.zaltren.firmabridge.config.FirmaBridgeConfig;
 import net.zaltren.firmabridge.integration.TFCStoneTypeHandler;
 import net.zaltren.firmabridge.registry.RockRegistry;
 import net.zaltren.firmabridge.worldgen.GTVeinPatcher;
+import net.zaltren.firmabridge.worldgen.VillageWorldGen;
 
 public class CommonProxy {
 
@@ -24,6 +26,10 @@ public class CommonProxy {
     public void init(FMLInitializationEvent event) {
         if (FirmaBridgeConfig.enableGTIntegration) {
             MaterialBridgeHandler.register();
+        }
+        if (FirmaBridgeConfig.enableVillages) {
+            GameRegistry.registerWorldGenerator(new VillageWorldGen(), 0);
+            FirmaBridge.LOGGER.info("FirmaBridge: village world generator registered.");
         }
     }